feat: Add minimal required functions to display the live dashboard
Implement all functions necessary to start a bokeh server which is able
to display live power data as a stacked area visualizaiton. At the
moment dummy data is displayed, since neither synthetic nor real data is
available. This needs to be changed in a future commit.

This commit closes issue #2